Upgrade Gradle, Gradle plugin, compile SDK version, target SDK version and libraries
I wanted to do a minor change but I could not run the tests because the old version of Gradle this library uses is not compatible with the latest NDK, which has dropped support for some ABIs. Here I update out-of-date things to be able to compile and run the tests.

Since I had to update the tests for them to run, I have migrated them to AndroidX.