Add option to change the text color of coach marks
We add the option at the CoachMark level so all coach marks can benefit from it. Note that HighlightCoachMarks have no text, but it is possible to set a text color in their builder which has no effect, the same way that it is possible to set a text in their builder which has no effect.